M3 2114 update: Add reduxed images to LinodeRebuild (#4305)
* Add reduxed images to Linoderebuild

- Add withImages to Linoderebuild
- Use ImageAndPassword in LinodeRebuild

* Fix tests

* Fix error handling
