+++ src/features/linodes/LinodesDetail/LinodeRebuild/LinodeRebuild.tsx
@@ -2,7 +2,7 @@ import { InjectedNotistackProps, withSnackbar } from 'notistack';
 import { isNil, pathOr } from 'ramda';
 import * as React from 'react';
 import { compose } from 'recompose';
-import AccessPanel, { UserSSHKeyObject } from 'src/components/AccessPanel';
+import { UserSSHKeyObject } from 'src/components/AccessPanel';
 import ActionsPanel from 'src/components/ActionsPanel';
 import Button from 'src/components/Button';
 
@@ -11,17 +11,15 @@ import { StyleRulesCallback, withStyles, WithStyles } from 'src/components/core/
 import Typography from 'src/components/core/Typography';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { Item } from 'src/components/EnhancedSelect/Select';
-import ErrorState from 'src/components/ErrorState';
 import SectionErrorBoundary from 'src/components/SectionErrorBoundary';
 import { resetEventsPolling } from 'src/events';
 import userSSHKeyHoc from 'src/features/linodes/userSSHKeyHoc';
-import { getImages } from 'src/services/images';
 import { rebuildLinode } from 'src/services/linodes';
 import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import { withLinode } from '../context';
 
-import { ImageSelect } from 'src/features/Images';
+import ImageAndPassword from '../LinodeSettings/ImageAndPassword';
 
 type ClassNames = 'root'
  | 'title'
@@ -59,8 +57,6 @@ interface ContextProps {
 }
 
 interface State {
-  images: Linode.Image[];
-  imagesError?: string;
   errors?: Linode.ApiFieldError[];
   selected?: string;
   password?: string;
@@ -74,16 +70,7 @@ type CombinedProps =
 
 class LinodeRebuild extends React.Component<CombinedProps, State> {
   state: State = {
-    images: [],
-  }
-
-  componentDidMount() {
-    getImages().then(response => {
-      this.setState({ images: response.data });
-    })
-      .catch(error => {
-        this.setState({ imagesError: error }); // @todo check typing
-      })
+    errors: []
   }
 
   handleImageSelect = (selectedItem: Item<string> | null) => {
@@ -91,7 +78,7 @@ class LinodeRebuild extends React.Component<CombinedProps, State> {
       this.setState({ selected: undefined });
       return;
     }
-    this.setState({ selected: selectedItem!.value });
+    this.setState({ selected: selectedItem!.value, errors: [] });
   }
 
   onSubmit = () => {
@@ -148,19 +135,10 @@ class LinodeRebuild extends React.Component<CombinedProps, State> {
 
   render() {
     const { classes, linodeLabel, userSSHKeys } = this.props;
-    const { errors, images, imagesError } = this.state;
-
-    if (imagesError) {
-      return (
-        <React.Fragment>
-          <DocumentTitleSegment segment={`${linodeLabel} - Rebuild`} />
-          <ErrorState errorText="There was an error retrieving images information." />
-        </React.Fragment>
-      );
-    }
+    const { errors } = this.state;
 
     const getErrorFor = getAPIErrorFor({}, errors);
-    const imageError = getErrorFor('image');
+    const imageFieldError = getErrorFor('image'); // @todo move out of render and use utility functions
     const passwordError = getErrorFor('password');
 
     return (
@@ -181,19 +159,14 @@ class LinodeRebuild extends React.Component<CombinedProps, State> {
             either restore from a backup or start over with a fresh Linux
             distribution. Rebuilding will destroy all data.
           </Typography>
-          <ImageSelect
-            images={images}
-            imageError={imageError}
-            onSelect={this.handleImageSelect}
-          />
-          <AccessPanel
-            noPadding
+          <ImageAndPassword
+            imageFieldError={imageFieldError}
+            onImageChange={this.handleImageSelect}
+            onPasswordChange={this.onPasswordChange}
             password={this.state.password || ''}
-            handleChange={this.onPasswordChange}
-            error={passwordError}
-            users={userSSHKeys.length > 0 ? userSSHKeys : []}
+            passwordError={passwordError}
+            userSSHKeys={userSSHKeys.length > 0 ? userSSHKeys : []}
           />
-
           <ActionsPanel>
             <Button
               type="secondary"
@@ -222,7 +195,7 @@ export default compose<CombinedProps, Props>(
   SectionErrorBoundary,
   styled,
   userSSHKeyHoc,
-  withSnackbar
+  withSnackbar,
 )(LinodeRebuild);
