+++ src/features/linodes/LinodesDetail/LinodeSettings/LinodeDiskDrawer.test.tsx
@@ -26,8 +26,6 @@ const props = {
   label: "This drawer",
   filesystem: "ext4",
   size: 50,
-  imagesLoading: false,
-  imageError: undefined
 }
 
 const component = shallow(<LinodeDiskDrawer {...props} />)
