+++ src/features/linodes/LinodesDetail/LinodeSettings/LinodeDiskDrawer.tsx
@@ -14,7 +14,6 @@ import Grid from 'src/components/Grid';
 import ModeSelect, { Mode } from 'src/components/ModeSelect';
 import Notice from 'src/components/Notice';
 import TextField from 'src/components/TextField';
-import withImages from 'src/containers/withImages.container';
 import getAPIErrorsFor from 'src/utilities/getAPIErrorFor';
 
 import ImageAndPassword from './ImageAndPassword';
@@ -41,12 +40,6 @@ interface EditableFields {
   userSSHKeys?: UserSSHKeyObject[];
 }
 
-interface WithImages {
-  images?: Linode.Image[];
-  imagesLoading: boolean;
-  imageError?: string;
-}
-
 interface Props extends EditableFields {
   mode: 'create' | 'rename' | 'resize';
   open: boolean;
@@ -69,7 +62,7 @@ interface State {
   selectedMode: diskMode;
 }
 
-type CombinedProps = Props & WithImages & WithStyles<ClassNames>;
+type CombinedProps = Props & WithStyles<ClassNames>;
 
 export const modes = {
   EMPTY: 'create_empty' as diskMode,
@@ -205,8 +198,6 @@ export class LinodeDiskDrawer extends React.Component<CombinedProps, State> {
     const {
       open,
       mode,
-      images,
-      imageError,
       onSubmit,
       submitting,
       onClose,
@@ -218,6 +209,7 @@ export class LinodeDiskDrawer extends React.Component<CombinedProps, State> {
 
     const generalError = this.getErrors('none');
     const passwordError = this.getErrors('root_pass');
+    const imageFieldError = this.getErrors('image');
 
     return (
       <Drawer title={LinodeDiskDrawer.getTitle(mode)} open={open} onClose={onClose}>
@@ -235,9 +227,8 @@ export class LinodeDiskDrawer extends React.Component<CombinedProps, State> {
             {selectedMode === modes.EMPTY && <this.filesystemField /> }
             {selectedMode === modes.IMAGE &&
               <ImageAndPassword
-                images={images || []}
-                imageError={imageError}
                 onImageChange={this.onImageChange}
+                imageFieldError={imageFieldError}
                 password={password || ''}
                 passwordError={passwordError}
                 onPasswordChange={this.props.onPasswordChange}
@@ -266,8 +257,6 @@ const styled = withStyles(styles);
 
 const enhanced = compose<CombinedProps, Props>(
   styled,
-  withImages((ownProps, images, imagesLoading, imageError) =>
-    ({ ...ownProps, images, imagesLoading, imageError })),
 )(LinodeDiskDrawer);
 
 export default enhanced;
